**Jira Ticket:** [SKYVERA-3179](https://jira.devfactory.com/browse/SKYVERA-3179) (#1857)
**TeamCity:** [TC Build Link]()
**RCA:** This issue is only replicating in one case when we remove unique permission from Item level.
There is a check in code which calls while creating or Updating project record. On the new record, it creates the groups on update it first checks if unique permission is setup on not. in the case of not-unique permissions, it re-creates the same project group with some incremental index value in SharePoint, and there was no code to check if the item already has groups attached or not.

**Solution:** While creating group we have added code to check if current item has already all three groups ("Owner", "Member", "Visitor"). If they exist then we are skipping the further processing of group creation.