https://jira.devfactory.com/browse/SKYVERA-3737 (#1869)
Cause: We had an issue while setting up RR from Schedule page. We can run RR with two ways one with Run directly and second is by schedule time. If we do both at the same time it creates a duplicate job because they both use different schedule type.

Solution: Now both jobs are using the same schedule type the only difference is with runtime value which will be nullable in the direct run case and when we schedule, it will update the same job and we will have only one job at a time.