Jira Ticket: https://jira.devfactory.com/browse/SKYVERA-3309 (#1893)
* Jira Ticket: https://jira.devfactory.com/browse/SKYVERA-3309

Recently we were having multiple issues of OutOfMemory exception while reporting refresh.
RCA: We are loading millions of rows in the data-table which sometimes throws OutOfMemory exception as per The load on the Server.

Solution: Now, instead of loading all data at the same time we are using batches. The default will be the same as it is now for all customers. But we can enable it for customers, those are having huge data for timesheets and it needs to run the following script and it is only applicated after this ticket is RELEASED.

* code changes

* code changes

* test changes

* removed comments

removed comments

* https://jira.devfactory.com/browse/SKYVERA-3309
Code review changes