**Jira Ticket:** [SKYVERA-4628](https://jira.devfactory.com/browse/SKYVERA-4628) **TeamCity:** [TC Build Link]()
**Changes proposed in the pull request:**

RCA: We have issues with Hidden Columns in List mapping they never considered in mapping but in Resource List case EXTID is hidden field but it is a link between Reporting and PFE resource entry which was broken for customer because of Remapping.The resource mapping is default provided in the template with EXTID. When Customer did the remapping of Resource List the column got delete from RPTColumns and after running cleanup data was null for EXTID column all Rows. it has also affected Resource Planner and Analyzer.

Solution: We cannot expose all Hidden columns because some of them used for Sharepoint purpose only so we added a check for EXTID column for Resource List mapping only if it does'not exists we are adding it in code.