**Jira Ticket:** [SKYVERA-4822](https://jira.devfactory.com/browse/SKYVERA-4822)
**Changes proposed in the pull request:**

Cause: We had an issue with one of the WorkHour events where we set EXTID for newly created WorkHours record. The EXT Id is the main link between resources and ResourcePlan available hours.
The code was setting EDTID for the existing record but not for a new record.
Solution: The code has been improved and it is working fine now.