Ticket : SKYVERA-6641
Cause :  we have issue with one of the contraints of TSMeta table. The table is used to store meta data of Lists for which we are submitting timesheets and We can setup columns which we want to store in TSMeta. In this case we had common columns in  two lists which were causing unique constraint error.The new adding constraint code was already there in SQL file but was not properly handled.

Solution :  The SQL code has been improved and tested on Dev machine as well as on SBC it is working fine